MAINT: move form factor tests to separate file
